feat: saveMovieButton change style on click
